+++ src/main/java/roomescape/domain/MemberName.java
@@ -1,21 +1,21 @@
-package roomescape.domain.vo;
+package roomescape.domain;
 
 import io.micrometer.common.util.StringUtils;
 import java.util.Objects;
 
-public class MemberRole {
+public class MemberName {
+    private static final int MAX_NAME_SIZE = 10;
 
-    private static final int MAX_ROLE_SIZE = 10;
     private final String value;
 
-    public MemberRole(String value) {
+    public MemberName(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (StringUtils.isBlank(value) || value.length() > MAX_ROLE_SIZE) {
-            throw new IllegalArgumentException("사용자의 권한은 1글자 이상 10글자 이하 입니다.");
+        if (StringUtils.isBlank(value) || value.length() > MAX_NAME_SIZE) {
+            throw new IllegalArgumentException(value + " 사용자 이름 1글자 이상 10글자 이하를 입력해주세요.");
         }
     }
 
@@ -31,7 +31,7 @@ public class MemberRole {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MemberRole that = (MemberRole) o;
+        MemberName that = (MemberName) o;
         return Objects.equals(value, that.value);
     }
