+++ src/main/java/roomescape/domain/MemberPassword.java
@@ -1,21 +1,20 @@
-package roomescape.domain.vo;
+package roomescape.domain;
 
 import io.micrometer.common.util.StringUtils;
 import java.util.Objects;
 
-public class MemberName {
-    private static final int MAX_NAME_SIZE = 10;
+public class MemberPassword {
 
     private final String value;
 
-    public MemberName(String value) {
+    public MemberPassword(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (StringUtils.isBlank(value) || value.length() > MAX_NAME_SIZE) {
-            throw new IllegalArgumentException(value + " 사용자 이름 1글자 이상 10글자 이하를 입력해주세요.");
+        if (StringUtils.isBlank(value)) {
+            throw new IllegalArgumentException("비밀번호는 필수입니다.");
         }
     }
 
@@ -24,14 +23,14 @@ public class MemberName {
     }
 
     @Override
-    public boolean equals(final Object o) {
+    public boolean equals(Object o) {
         if (this == o) {
             return true;
         }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MemberName that = (MemberName) o;
+        MemberPassword that = (MemberPassword) o;
         return Objects.equals(value, that.value);
     }
