+++ src/main/java/roomescape/domain/MemberRole.java
@@ -1,20 +1,21 @@
-package roomescape.domain.vo;
+package roomescape.domain;
 
 import io.micrometer.common.util.StringUtils;
 import java.util.Objects;
 
-public class MemberPassword {
+public class MemberRole {
 
+    private static final int MAX_ROLE_SIZE = 10;
     private final String value;
 
-    public MemberPassword(String value) {
+    public MemberRole(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (StringUtils.isBlank(value)) {
-            throw new IllegalArgumentException("비밀번호는 필수입니다.");
+        if (StringUtils.isBlank(value) || value.length() > MAX_ROLE_SIZE) {
+            throw new IllegalArgumentException("사용자의 권한은 1글자 이상 10글자 이하 입니다.");
         }
     }
 
@@ -23,14 +24,14 @@ public class MemberPassword {
     }
 
     @Override
-    public boolean equals(Object o) {
+    public boolean equals(final Object o) {
         if (this == o) {
             return true;
         }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MemberPassword that = (MemberPassword) o;
+        MemberRole that = (MemberRole) o;
         return Objects.equals(value, that.value);
     }
