+++ src/main/java/roomescape/domain/ReservationDate.java
@@ -1,4 +1,4 @@
-package roomescape.domain.vo;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
