+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,8 +7,8 @@ import java.util.Objects;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationRepository;
+import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
-import roomescape.domain.vo.ReservationTime;
 import roomescape.exception.ReservationExistsException;
 import roomescape.service.request.ReservationTimeAppRequest;
 import roomescape.service.response.BookableReservationTimeAppResponse;
