+++ src/test/java/roomescape/domain/MemberPasswordTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.vo;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,22 +7,21 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.junit.jupiter.params.provider.ValueSource;
 
-class MemberRoleTest {
-    @DisplayName("생성 테스트")
+class MemberPasswordTest {
+
+    @DisplayName("올바른 비밀 번호 형식이면 생성된다.")
     @Test
     void create() {
-        assertThatCode(() -> new MemberRole("ADMIN"))
+        assertThatCode(() -> new MemberPassword("abc1234"))
             .doesNotThrowAnyException();
     }
 
-    @DisplayName("빈 값이거나 10글자 이상이면 예외가 발생한다.")
+    @DisplayName("비밀 번호가 빈 값이면 예외가 발생한다.")
     @ParameterizedTest
     @NullAndEmptySource
-    @ValueSource(strings = {"11111111111"})
-    void create_Fail(String role) {
-        assertThatThrownBy(() -> new MemberRole(role))
+    void create_Fail(String input) {
+        assertThatThrownBy(() -> new MemberPassword(input))
             .isInstanceOf(IllegalArgumentException.class);
     }
 }
