+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -23,10 +23,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.domain.MemberRepository;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationRepository;
+import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
 import roomescape.domain.Theme;
 import roomescape.domain.ThemeRepository;
-import roomescape.domain.vo.ReservationTime;
 import roomescape.exception.PastReservationException;
 import roomescape.service.request.ReservationAppRequest;
 import roomescape.service.response.ReservationAppResponse;
